Updated Building System
If the tile the player wants to build an object on is next to an object thats already been built then the player can build an object there. As well as the building system now checks if the object being built is overlapping an object already built and if it is then the player can't build.